Rename gravity gradient files
